coloring in compilation buffer
for some reason unbeknownst to me, emacs standard compilation cannot
handle properly ascii color escape codes (probably because there are
codes that are only for the terminal? idk). either way, adding that
hook is kind of a nasty way to fix it, and i would like to do without
it but dont really know how right now.

will probably check it later.