automatically find entry point
now if you comment `kind-mode-entry-point` on the term name definition
line it will automatically use that when you run `kind-run-term` on
the buffer.

note that the term definition should only span one line (so that i can
create a regex to match it without using complex parsers and stuff).

TODO: make this a proper multi-line search.